Attempt at file write included
I've included how I think the file write section of the js part of the task should work. I'm running into an error that I can't resolve. I think it may be related to my dev environment and that it requires NodeJS. I've tried npm install and using npm run dev and that doesn't seem to fix it. There doesn't seem to be a way to write files with vanilla js, I did try using File() but couldn't make any headway with it.